added monitor, file skip

+++ copy_rename_files.py
@@ -10,6 +10,7 @@ sys.path.append(os.path.join(cd,'utils'))
 import re
 import glob
 import shutil
+import numpy as np
 
 #%% Inputs
 source='C:/Users/SLETIZIA/Box/wfip3_barge_campbell_raw/*dat'
@@ -20,6 +21,7 @@ regex_out='barg.met.z01.00.{DD}.{HH}00.dat'
 
 #%% Initialization
 files=sorted(glob.glob(source))
+ctr=0
 
 #%% Main
 for f in files:
@@ -27,5 +29,9 @@ for f in files:
     new_name=regex_out
     for group_name in match.groupdict():
         new_name = new_name.replace('{'+group_name+'}',match.group(group_name))
-    shutil.copy(f,os.path.join(destination,new_name))
+    if os.path.exists(os.path.join(destination,new_name))==False:
+        shutil.copy(f,os.path.join(destination,new_name))
+    
+    print(str(np.round(ctr/len(files)*100,2))+ '% done')
+    ctr+=1
